+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `nitrokey.nk3.secrets_app`: Fix discarded authentication when using `SecretsApp` over CCID
 - Bump minimum Python version to 3.11.
+- `nitrokey.trussed`: Add the `inner_checksum` field to `FirmwareMetadata` with the checksum of the unsigned firmware image.
 
 [All Changes](https://github.com/Nitrokey/nitrokey-sdk-py/compare/v0.5.0-rc.2...HEAD)